Move SessionRefreshModal into its own file
